+++ examples/full-nixos/arion-compose.nix
@@ -1,4 +1,5 @@
 {
+  project.name = "full-nixos";
   services.webserver = { pkgs, lib, ... }: {
     nixos.useSystemd = true;
     nixos.configuration.boot.tmpOnTmpfs = true;