+++ src/nix/modules.nix
@@ -2,6 +2,7 @@
     ./modules/composition/docker-compose.nix
     ./modules/composition/host-environment.nix
     ./modules/composition/images.nix
+    ./modules/composition/networks.nix
     ./modules/composition/service-info.nix
     ./modules/composition/composition.nix
 ]